// Source/SelfService/Web/apis/integrations/generated/models/IonConfigRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aigonix.Bridge.M3
 * Bridge API - made for Aigonix Studio
 *
 * The version of the OpenAPI document: 0.0.1.333
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface IonConfigRequest
 */
export interface IonConfigRequest {
    /**
     * 
     * @type {string}
     * @memberof IonConfigRequest
     */
    iu: string;
    /**
     * 
     * @type {string}
     * @memberof IonConfigRequest
     */
    pu: string;
    /**
     * 
     * @type {string}
     * @memberof IonConfigRequest
     */
    ot: string;
    /**
     * 
     * @type {string}
     * @memberof IonConfigRequest
     */
    saak: string;
    /**
     * 
     * @type {string}
     * @memberof IonConfigRequest
     */
    sask: string;
    /**
     * 
     * @type {string}
     * @memberof IonConfigRequest
     */
    ci: string;
    /**
     * 
     * @type {string}
     * @memberof IonConfigRequest
     */
    cs: string;
    /**
     * 
     * @type {string}
     * @memberof IonConfigRequest
     */
    ti?: string | null;
}

/**
 * Check if a given object implements the IonConfigRequest interface.
 */
export function instanceOfIonConfigRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "iu" in value;
    isInstance = isInstance && "pu" in value;
    isInstance = isInstance && "ot" in value;
    isInstance = isInstance && "saak" in value;
    isInstance = isInstance && "sask" in value;
    isInstance = isInstance && "ci" in value;
    isInstance = isInstance && "cs" in value;

    return isInstance;
}

export function IonConfigRequestFromJSON(json: any): IonConfigRequest {
    return IonConfigRequestFromJSONTyped(json, false);
}

export function IonConfigRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): IonConfigRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'iu': json['iu'],
        'pu': json['pu'],
        'ot': json['ot'],
        'saak': json['saak'],
        'sask': json['sask'],
        'ci': json['ci'],
        'cs': json['cs'],
        'ti': !exists(json, 'ti') ? undefined : json['ti'],
    };
}

export function IonConfigRequestToJSON(value?: IonConfigRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'iu': value.iu,
        'pu': value.pu,
        'ot': value.ot,
        'saak': value.saak,
        'sask': value.sask,
        'ci': value.ci,
        'cs': value.cs,
        'ti': value.ti,
    };
}